requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
crypto==1.4.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
Naked==0.1.32
packaging==24.1
pillow==10.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
